+++ apps/web/src/features/auth/sign-in/index.tsx
@@ -10,7 +10,7 @@ import { AuthLayout } from '../auth-layout'
 import { UserAuthForm } from './components/user-auth-form'
 
 export function SignIn() {
-  const { redirect } = useSearch({ from: '/(auth)/sign-in' })
+  const { redirect } = useSearch({ from: '/(auth)/login' })
 
   return (
     <AuthLayout>